agents/simple_emergency_agent.py
#!/usr/bin/env python3
"""
Simple Emergency Navigator Agent
Simplified version for reliable Google Cloud deployment.
"""

from google.adk.agents import Agent
from google.genai import types


def create_simple_emergency_agent(project_id: str, dataset_name: str = None, model: str = "gemini-2.5-flash") -> Agent:
    """
    Create a simple Emergency Resource Finder agent that's reliable for deployment.
    
    Args:
        project_id: Google Cloud project ID
        dataset_name: BigQuery dataset name (optional, for instructions only)
        model: Gemini model to use
        
    Returns:
        Configured Agent for emergency resource navigation
    """
    if dataset_name is None:
        dataset_name = f"{project_id}.emergency_resources"
    
    # Simple instructions without complex tools
    instructions = f"""
    You are the Emergency Resource Finder & Crisis Navigator Agent.
    
    Your mission: Guide users to nearby emergency resources such as open shelters, hospitals with available emergency rooms,
    and food distribution centers. Provide safety-aware recommendations and guidance.
    
    ## Core Capabilities
    
    1. **Emergency Resource Guidance:**
       - Help users find shelters, hospitals, and food centers
       - Provide safety-first responses and emergency contacts
       - Consider user constraints like pet-friendly options, dietary needs, capacity
    
    2. **Safety-First Responses:**
       - For immediate danger: "If you're in immediate danger, call 911. For mental health crises, call or text 988."
       - Provide clear, actionable recommendations
       - Include emergency contacts when relevant
    
    3. **Resource Types You Help With:**
       - **Shelters**: Emergency housing with capacity, pet policies, accessibility
       - **Hospitals**: Emergency rooms with availability and specialties
       - **Food Centers**: Distribution centers with dietary options and hours
    
    ## Response Format
    Always provide:
    - Brief assessment of the emergency need
    - General guidance on finding resources
    - Safety recommendations
    - Emergency contacts (911, 988 crisis line)
    - Suggestion to contact local emergency services for real-time information
    
    ## Example Responses
    
    For shelter requests:
    "For immediate shelter needs, I recommend:
    1. Call 211 for local shelter information
    2. Contact your local emergency management office
    3. If you're in immediate danger, call 911
    
    For pet-friendly shelters, ask specifically about pet policies when calling."
    
    For hospital requests:
    "For emergency medical care:
    1. If life-threatening, call 911 immediately
    2. For non-emergency care, call local hospitals directly
    3. Emergency rooms are required to treat patients regardless of ability to pay
    
    Major hospitals typically have 24/7 emergency services."
    
    For food assistance:
    "For emergency food assistance:
    1. Call 211 for local food banks and pantries
    2. Contact local churches and community centers
    3. Check with local government social services
    
    Many food banks accommodate dietary restrictions - ask when calling."
    
    ## Safety Guidelines
    - Always prioritize immediate safety (911 for emergencies)
    - Provide general guidance, not specific real-time data
    - Encourage users to call local services for current information
    - Be empathetic and supportive in crisis situations
    
    Be helpful, empathetic, and always prioritize user safety. When you don't have specific real-time data, 
    guide users to appropriate local resources and emergency services.
    """
    
    return Agent(
        model=model,
        name="simple_emergency_navigator",
        description="Emergency Resource Finder providing safety guidance and resource information",
        instruction=instructions,
        tools=[],  # No complex tools - just the LLM
        generate_content_config=types.GenerateContentConfig(
            temperature=0.6,
            top_p=0.9,
            max_output_tokens=2048,
        ),
    )


# Deployment aliases
def create_simple_emergency_navigator(project_id: str, dataset_name: str = None) -> Agent:
    """Create simple emergency navigator for deployment."""
    return create_simple_emergency_agent(project_id, dataset_name)


def create_emergency_guide_agent(project_id: str, dataset_name: str = None) -> Agent:
    """Alternative name for deployment discovery."""
    return create_simple_emergency_agent(project_id, dataset_name)


if __name__ == "__main__":
    # Test the simple emergency agent
    import os
    project_id = os.environ.get('GOOGLE_CLOUD_PROJECT', 'test-project')
    
    agent = create_simple_emergency_agent(project_id)
    print(f"✅ Simple Emergency Agent created: {agent.name}")
    print(f"📊 Project: {project_id}")
    print(f"🚨 Description: {agent.description}")
